% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prml_analysis.R
\name{dtgamma}
\alias{dtgamma}
\title{Calculate pdf of the truncated gamma}
\usage{
dtgamma(x_, shape_, rate_, a, b)
}
\arguments{
\item{x_}{A number, value.}

\item{shape_}{A number, shape of the truncated gamma.}

\item{rate_}{A number, rate of the truncated gamma.}

\item{a}{lower bound of the range}

\item{b}{upper bound of the range}
}
\value{
A vector, pdf of \eqn{ Ga_{[a,b]}(x\_; shape\_,rate\_) }
}
\description{
Calculate pdf of the truncated gamma
}
