% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prml_analysis.R
\name{logmean}
\alias{logmean}
\title{Calculate the \eqn{\log(\text{mean}(\text{exp}(x)))}}
\usage{
logmean(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
the log(mean(exp(x)))
}
\description{
Calculate the \eqn{\log(\text{mean}(\text{exp}(x)))}
}
