% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prml_analysis.R
\name{logsum}
\alias{logsum}
\title{Calculate the log(sum(exp(x)))}
\usage{
logsum(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A number, the \eqn{ log(sum(exp(x))) }
}
\description{
Calculate the log(sum(exp(x)))
}
