% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prml_analysis.R
\name{prml.mix}
\alias{prml.mix}
\title{PRML classifier: Posterior probability under each Poisson mixtures hypotheses.}
\usage{
prml.mix(xs_bn, xs_a, xs_b, mu_l = "min", mu_u = "max", e = 0,
  r_a = 0.5, s_a = 2e-10, r_b = 0.5, s_b = 2e-10, n_gq = 20,
  n_per = 100)
}
\arguments{
\item{xs_bn}{A vector. Spike counts of repeated dual-stimuli trial data AB.}

\item{xs_a}{A vector. Spike counts of repeated single-stimulus trial data A.}

\item{xs_b}{A vector. Spike counts of repeated single-stimulus trial data B.}

\item{mu_l}{A number. Lower bound of spike counts. "min" by default. Indicating \equ{max(0,\underset{j=A,B,AB}{\min}(\min(Y_j)-2\text{std}(Y_j)))}}

\item{mu_u}{A number. Upper bound of spike counts. "max" by default. Indicating \equ{\underset{j=A,B,AB}{\max}(\max(Y_j)+2\text{std}(Y_j))}}

\item{e}{A number. 0 by default. Shringkage on the domain and meansurement of mixing density f under the Intermediate and Mixture hypothese.}

\item{r_a}{A number. The parameter in gamma prior of spike rate mu_A. rate. Jeffereys' prior by default.}

\item{s_a}{A number. The parameter in gamma prior of spike rate mu_A. shape. Jeffereys' prior by default.}

\item{r_b}{A number. The parameter in gamma prior of spike rate mu_B. rate. Jeffereys' prior by default.}

\item{s_b}{A number. The parameter in gamma prior of spike rate mu_B. shape. Jeffereys' prior by default.}

\item{n_gq}{A number. 20 by default. Number of grids in Gaussion quadrature.}

\item{n_per}{A number. 100 by default. Permutation of likihood estimation to obtain the order-invariant estimator.}
}
\value{
A list. post.prob: posterior probabilities under Mixture, Intermediate, Outside, Single hypotheses. win.model: the model has largest post.prob.
}
\description{
PRML classifier: Posterior probability under each Poisson mixtures hypotheses.
}
