% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prml_analysis.R
\name{prml_filter}
\alias{prml_filter}
\title{PRML filter: Bayes factor of Poisson versus Poisson mixtures.}
\usage{
prml_filter(xs_bn, n_gq = 20, n_per = 100, alpha = 0.5)
}
\arguments{
\item{xs_bn}{A vector. Spike counts of repeated single-stimulus trial data.}

\item{n_gq}{A number. 20 by default. Number of grids in Gaussion quadrature.}

\item{n_per}{A number. 100 by default. Permutation of likihood estimation to obtain the order-invariant estimator.}

\item{alpha}{0.5 by default. The range of the spike counts estimator \equ{[Y_{0.25}-\code{alpha}IQR,Y_{0.75}+\code{alpha}IQR]}}
}
\value{
A number. Bayes factor of Poisson versus Poisson mixtures by PRML algorithm.
}
\description{
PRML filter: Bayes factor of Poisson versus Poisson mixtures.
}
