% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prml_analysis.R
\name{dtgamma}
\alias{dtgamma}
\title{Calculate pdf of the truncated gamma \equ{Ga_{[\code{a},\code{b}]}(\code{shape_},\code{rate_})}}
\usage{
dtgamma(x_, shape_, rate_, a, b)
}
\arguments{
\item{x_}{A number, value.}

\item{shape_}{A number, shape of the truncated gamma.}

\item{rate_}{A number, rate of the truncated gamma.}

\item{a}{lower bound of the range}

\item{b}{upper bound of the range}
}
\value{
A vector, pdf of \equ{Ga_{[\code{a},\code{b}]}(\code{x}; \code{shape_},\code{rate_})}
}
\description{
Calculate pdf of the truncated gamma \equ{Ga_{[\code{a},\code{b}]}(\code{shape_},\code{rate_})}
}
