% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prml_analysis.R
\name{logmean}
\alias{logmean}
\title{Calculate the log(mean(exp(x)))}
\usage{
logmean(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A number, the \eqn{ \text{log}(\text{mean}(\text{exp}(x))) }
}
\description{
Calculate the log(mean(exp(x)))
}
