% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prml_analysis.R
\name{prml_tests_f}
\alias{prml_tests_f}
\title{PRML classifier together with pdf of mixing density.}
\usage{
prml_tests_f(xA, xB, xAB, labels = c("A", "B", "AB"),
  remove.zeros = FALSE, mu_l = "min", mu_u = "max", e = 0,
  gamma.pars = c(0.5, 2e-10), n_gq = 20, n_mu = 100, n_per = 100,
  alpha = 0.5)
}
\arguments{
\item{mu_u}{A number. 100 by default. The number of grids used to represent the pdf of f.}
}
\value{
A list.
\describe{
  \item{out1}{Result of \code{prml_tests}}
  \item{out2}{pdf of mixing density f under Mixture, Intermediate, OutsideA, OutsideB hypotheses}
}
}
\description{
PRML classifier together with pdf of mixing density.
}
\seealso{
\code{\link{prml_tests}}
}
